Fixed misc issues
When I downloaded Angular I got version 1.3 and almost everything
works. The only issue I found was when I add added a birthday field the
value exists only until the 1 secon timeout and goes away. The other
fields seemed to work.

I had watched the videos and found them to be helpful. Then I
downloaded the project and had a little trouble getting started. I
found the information that I needed in lesson 2 so I added it to the
readme. There also a few express warnings that I fixed.